+++ tests/test_selfhost_lexer_gen.c
@@ -328,7 +328,7 @@ Lexer lx_skip_ws_and_comments(Lexer * lx __attribute__((unused))) {
     int len __attribute__((unused)) = (lx)->len;
     Lexer __loop_result_2 __attribute__((unused));
     int pos = (lx)->pos;
-    double line = (lx)->line;
+    int line = (lx)->line;
     int in_comment = 0;
     while (1) {
         if ((pos >= len)) {
@@ -339,7 +339,7 @@ Lexer lx_skip_ws_and_comments(Lexer * lx __attribute__((unused))) {
         if (in_comment) {
         if ((c == 10)) {
         int __recur_tmp_0 = pos;
-        double __recur_tmp_1 = line;
+        int __recur_tmp_1 = line;
         int __recur_tmp_2 = 0;
         pos = __recur_tmp_0;
         line = __recur_tmp_1;
@@ -347,7 +347,7 @@ Lexer lx_skip_ws_and_comments(Lexer * lx __attribute__((unused))) {
         continue;
         } else {
         int __recur_tmp_0 = (pos + 1);
-        double __recur_tmp_1 = line;
+        int __recur_tmp_1 = line;
         int __recur_tmp_2 = 1;
         pos = __recur_tmp_0;
         line = __recur_tmp_1;
@@ -357,7 +357,7 @@ Lexer lx_skip_ws_and_comments(Lexer * lx __attribute__((unused))) {
         } else {
         if (lx_is_whitespace_(c)) {
         int __recur_tmp_0 = (pos + 1);
-        double __recur_tmp_1 = ((c == 10) ? (line + 1) : line);
+        int __recur_tmp_1 = ((c == 10) ? (line + 1) : line);
         int __recur_tmp_2 = 0;
         pos = __recur_tmp_0;
         line = __recur_tmp_1;
@@ -366,7 +366,7 @@ Lexer lx_skip_ws_and_comments(Lexer * lx __attribute__((unused))) {
         } else {
         if ((c == 59)) {
         int __recur_tmp_0 = (pos + 1);
-        double __recur_tmp_1 = line;
+        int __recur_tmp_1 = line;
         int __recur_tmp_2 = 1;
         pos = __recur_tmp_0;
         line = __recur_tmp_1;
@@ -507,7 +507,7 @@ Result lex_string(Lexer * lx __attribute__((unused)), Arena *dest __attribute__(
     int len __attribute__((unused)) = (lx)->len;
     Result __loop_result_6 __attribute__((unused));
     int pos = ((lx)->pos + 1);
-    double line = start_line;
+    int line = start_line;
     char * acc = "";
     while (1) {
         if ((pos >= len)) {
@@ -527,7 +527,7 @@ Result lex_string(Lexer * lx __attribute__((unused)), Arena *dest __attribute__(
         int e __attribute__((unused)) = char_at(src, (pos + 1));
         int decoded __attribute__((unused)) = ((e == 110) ? 10 : ((e == 116) ? 9 : ((e == 34) ? 34 : ((e == 92) ? 92 : e))));
         int __recur_tmp_0 = (pos + 2);
-        double __recur_tmp_1 = line;
+        int __recur_tmp_1 = line;
         char * __recur_tmp_2 = concat(acc, char_from_code(decoded, dest), dest);
         pos = __recur_tmp_0;
         line = __recur_tmp_1;
@@ -536,7 +536,7 @@ Result lex_string(Lexer * lx __attribute__((unused)), Arena *dest __attribute__(
         }
         } else {
         int __recur_tmp_0 = (pos + 1);
-        double __recur_tmp_1 = ((c == 10) ? (line + 1) : line);
+        int __recur_tmp_1 = ((c == 10) ? (line + 1) : line);
         char * __recur_tmp_2 = concat(acc, char_from_code(c, dest), dest);
         pos = __recur_tmp_0;
         line = __recur_tmp_1;
